backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
openai==1.6.1
PyPDF2==3.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
python-docx==0.8.11
httpx==0.25.2